requirements.txt
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
autoflake==2.2.1
billiard==4.2.0
black==24.1.1
celery==5.3.6
certifi==2024.2.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.2
fastapi-pagination==0.12.15
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
isort==5.13.2
kombu==5.3.5
Mako==1.3.2
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.5.1
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
pyflakes==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.26
SQLAlchemy-Utils==0.41.1
starlette==0.36.3
tenacity==8.2.3
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
ujson==5.9.0
uvicorn==0.27.1
vine==5.1.0
wcwidth==0.2.13
